pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "expensely"
version = "2020.1"
description = "Expensely expense management"
readme = "README.rst"
repository = "https://github.com/inducer/expensely"
authors = ["Andreas Kloeckner <inform@tiker.net>"]
license = "MIT"
packages = [
    { include = "expenses" },
    { include = "expensely" },
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
# Automatically renders Django forms in a pretty, Bootstrap-compatible way.
django-crispy-forms = ">=2.3"
crispy-bootstrap3 = "^2024.1"
pytz = ">=2020.1"
# A date picker widget - https://github.com/tutorcruncher/django-bootstrap3-datetimepicker
django-bootstrap3-datetimepicker-2 = "^2.8.2"

psycopg2 = { version = "^2.8.5", optional = true }

[tool.poetry.group.dev.dependencies]
ruff = "^0.9"

# enable with "-E postgres"
[tool.poetry.extras]
postgres = ["psycopg2"]

[tool.ruff]
preview = true
target-version = "py310"
exclude = ["contrib/jupyterlite"]

[tool.ruff.lint]
extend-select = [
    "B",   # flake8-bugbear
    "C",   # flake8-comprehensions
    "E",   # pycodestyle
    "F",   # pyflakes
    "I",   # flake8-isort
    "N",   # pep8-naming
    "NPY", # numpy
    "Q",   # flake8-quotes
    "W",   # pycodestyle
    "RUF",
    "UP",
    # "DJ",
]

extend-ignore = [
    "C90",  # McCabe complexity
    "E221", # multiple spaces before operator
    "E226", # missing whitespace around arithmetic operator
    "E241", # multiple spaces after comma
    "E242", # tab after comma
    "E402", # module level import not at the top of file
    "N818", # error suffix in exception names

    # TODO
    "B904", # raise in except needs from
    "B028", # stacklevel
    "RUF012", # mutable class atttributes
    "UP031", # %-format
]
allowed-confusables = ["‐", "–"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "double"
multiline-quotes = "double"

[tool.ruff.lint.isort]
combine-as-imports = true
lines-after-imports = 2

[tool.ruff.lint.per-file-ignores]
# mostly generated
"*/migrations/*.py" = ["Q", "E501", "RUF012", "F401", "I002", "UP"]

